Moves precisionWasLost to helper module
